Refactored, moved all tester-related things into the TestingAPI package
