+++ app/src/test/java/Bamboo/TestingAPI/Metrics.java
@@ -1,4 +1,4 @@
-package Bamboo.ExperimentationWinRates;
+package Bamboo.TestingAPI;
 
 public enum Metrics {
     WIN_RATE,
