+++ app/src/test/java/Bamboo/TestingAPI/Parameter.java
@@ -1,4 +1,4 @@
-package Bamboo.ExperimentationWinRates;
+package Bamboo.TestingAPI;
 
 public enum Parameter {
     DEPTH,
