+++ app/src/test/java/Bamboo/TestingAPI/TesterAgent.java
@@ -1,4 +1,4 @@
-package Bamboo.ExperimentationWinRates;
+package Bamboo.TestingAPI;
 
 public enum TesterAgent {
     AGENT_1,
