+++ app/src/test/java/Bamboo/TestingAPI/TesterAgentFactory.java
@@ -1,4 +1,4 @@
-package Bamboo.ExperimentationWinRates;
+package Bamboo.TestingAPI;
 
 import Bamboo.controller.Agent;
