+++ app/src/test/java/Bamboo/TestingAPI/Variable.java
@@ -1,4 +1,4 @@
-package Bamboo.ExperimentationWinRates;
+package Bamboo.TestingAPI;
 
 public enum Variable {
     GRID_SIZE,
